client: Add mutex round EGLSurface manipulation
The EGLSurface is manipulated from the render thread when we first
create a window with makeCurrent potentially after it's
invalidated but also directly on the main thread through QWaylandWindow::updateGeometry.

It we resize a window at the exact moment of rendering the first frame
we can concurrently enter updateSurface from two threads. We risk
creating a window twice or using the wrong size.

Pick-to: 6.8
Change-Id: I2e5c3e86d3147854c420a4fdd7afcb17ca020c36
Reviewed-by: Vlad Zahorodnii <vlad.zahorodnii@kde.org>