Compositor: don't give protocol error when re-showing window
The wayland protocol explicitly allows setting the same role multiple
times. Therefore, we cannot send a protocol error just because the
surface role is already set. In xdg_wm_base_get_xdg_surface we don't
know which role we are going to set, so we cannot do the proper test
there. If the roles do not match, the error will be caught by
QWaylandSurface::setRole(), so it is safe to remove this check.

Fixes: QTBUG-130652
Change-Id: Ic90d05d87f1553479a6a39ac41c3c91c38962bdb
Reviewed-by: Eskil Abrahamsen Blomfeldt <eskil.abrahamsen-blomfeldt@qt.io>