MockCompositor: Fix potential race condition
The list m_viewports is being accessed from both the main thread
and the dispatch thread. This has not been observed to cause any
errors, but is a potential source of test instability. Using a
DirectConnection makes sure that both appending and removal
happens in the same thread.

Task-number: QTBUG-130627
Pick-to: 6.8
Change-Id: Iee6941d8794119e679a485778f0be5c40df27069
Reviewed-by: David Edmundson <davidedmundson@kde.org>