Adwaita decorations: improve border painting
This is a backport of two changes pushed to QAdwaitaDecorations upstream
that further improve painting of border around the application window.
Using the 0.5px adjustment makes the border to look better as the line
better aligns. Also the separator between window and title doesn't need
to be painted separately. Co-authored with Lorenzo Bertini.

Pick-to: 6.9 6.8
Change-Id: Ib841073884dee58e599a173c259547416d2ed8ce
Reviewed-by: Eskil Abrahamsen Blomfeldt <eskil.abrahamsen-blomfeldt@qt.io>