client: Redo management of tablet object proxies
Since 5af836aea3bb91a9f388e565415dc33b2fde4577 tools and pads can sometimes
be parented to tablets. When a tablet is unplugged sometimes the remove
event for the tablet can be sent before the remove events for the pad.
Deleting the tablet will also delete the pad but not the pad proxy,
resulting in a crash when the pad remove event is received.
This moves destruction of the wayland proxies which makes everything much
simpler as not every location that deletes those objects needs to call
destroy and fixes the described problem.

Pick-to: 6.8
Change-Id: I0aaeda3d3996251e411229b5e97aa1ce0bce43ac
Reviewed-by: David Edmundson <davidedmundson@kde.org>