Remove qt-key extension
This was an extension to make sure we send the full key
event the client was running against a Qt Wayland
Compositor.

But the world has moved on and this is now handled through
official extensions. And we have to support multiple
third-party compositors, so any issues that occur with
key event handling should be handled on a more "official"
level.

Task-number: QTBUG-133223
Change-Id: I96cf394fa0e79eeaeba6dfec0c0fc81f5320fdd4
Reviewed-by: Vlad Zahorodnii <vlad.zahorodnii@kde.org>